# docs/openapi/autoflow-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoFlow API",
    "version": "v1",
    "description": "Contrato OpenAPI versionado da API AutoFlow. Exportado a partir do contrato SpringDoc da aplicação e complementado com parâmetros de execução, autenticação e exemplos em docs/openapi/README.md."
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "API local (backend)"
    },
    {
      "url": "http://localhost:4200/api",
      "description": "API local pelo proxy do frontend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orcamentos/{orcamentoId}": {
      "get": {
        "operationId": "consultarOrcamentoAutenticado",
        "summary": "Consultar orçamento autenticado",
        "tags": [
          "orçamentos"
        ],
        "description": "Consultar orçamento autenticado Permissão: ADMIN, ATENDENTE ou CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      }
    },
    "/orcamentos/{orcamentoId}/aprovar": {
      "post": {
        "operationId": "aprovarOrcamentoAutenticado",
        "summary": "Aprovar orçamento autenticado",
        "tags": [
          "orçamentos"
        ],
        "description": "Aprovar orçamento autenticado Permissão: ADMIN ou CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      }
    },
    "/orcamentos/{orcamentoId}/recusar": {
      "post": {
        "operationId": "recusarOrcamentoAutenticado",
        "summary": "Recusar orçamento autenticado",
        "tags": [
          "orçamentos"
        ],
        "description": "Recusar orçamento autenticado Permissão: ADMIN ou CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecusarOrcamentoRequest"
              }
            }
          }
        }
      }
    },
    "/orcamentos/{orcamentoId}/pdf": {
      "get": {
        "operationId": "baixarOrcamentoAutenticado",
        "summary": "Baixar PDF do orçamento autenticado",
        "tags": [
          "orçamentos"
        ],
        "description": "Baixar PDF do orçamento autenticado Permissão: ADMIN, ATENDENTE ou CLIENTE.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ]
      }
    },
    "/public/orcamentos/{orcamentoId}": {
      "get": {
        "operationId": "consultarOrcamentoPublico",
        "summary": "Consultar orçamento pelo link público",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "O token público é obrigatório, não é persistido em claro e deve ser enviado na query string.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "security": []
      }
    },
    "/public/orcamentos/{orcamentoId}/pdf": {
      "get": {
        "operationId": "baixarOrcamentoPublico",
        "summary": "Baixar PDF pelo token público",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "Baixar PDF pelo token público",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "security": []
      }
    },
    "/public/orcamentos/{orcamentoId}/aprovar": {
      "post": {
        "operationId": "aprovarOrcamentoPublico",
        "summary": "Aprovar orçamento pelo link público",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "A decisão é idempotente para o mesmo estado e não exige JWT.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AprovarOrcamentoRequest"
              }
            }
          }
        },
        "security": []
      }
    },
    "/public/orcamentos/{orcamentoId}/recusar": {
      "post": {
        "operationId": "recusarOrcamentoPublico",
        "summary": "Recusar orçamento pelo link público",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "A decisão é idempotente para o mesmo estado e não exige JWT.",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecusarOrcamentoRequest"
              }
            }
          }
        },
        "security": []
      }
    },
    "/public/orcamentos/{orcamentoId}/pdf/acompanhamento": {
      "get": {
        "operationId": "baixarOrcamentoAcompanhamento",
        "summary": "Baixar PDF pelo token de acompanhamento",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "Baixar PDF pelo token de acompanhamento",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "security": []
      }
    },
    "/public/orcamentos/{orcamentoId}/aprovar/acompanhamento": {
      "post": {
        "operationId": "aprovarOrcamentoAcompanhamento",
        "summary": "Aprovar orçamento pelo acompanhamento",
        "tags": [
          "orçamentos públicos"
        ],
        "description": "Aprovar orçamento pelo acompanhamento",
        "responses": {
          "200": {
            "description": "Operação concluída com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrcamentoResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "parameters": [
          {
            "name": "orcamentoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<public-token>"
          }
        ],
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token retornado por POST /auth/login no header Authorization: Bearer <token>."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida, JSON inválido, validação ou regra de negócio.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "JSON inválido"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "JWT ausente/inválido ou token público ausente, inválido ou expirado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Token inválido ou expirado"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Usuário autenticado sem a role necessária.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Acesso negado"
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "Recurso não encontrado"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string",
            "example": "Recurso não encontrado"
          },
          "campos": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AprovarOrcamentoRequest": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 120,
            "example": "Maria da Silva"
          }
        }
      },
      "RecusarOrcamentoRequest": {
        "type": "object",
        "properties": {
          "motivo": {
            "type": "string",
            "maxLength": 500,
            "example": "Valor acima do esperado"
          },
          "nome": {
            "type": "string",
            "maxLength": 120,
            "example": "Maria da Silva"
          }
        }
      },
...
      "OrcamentoServicoResponse": {
        "type": "object",
        "properties": {
          "servicoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Troca de óleo"
          },
          "valor": {
            "type": "number",
            "format": "double",
            "example": 150
          }
        }
      },
      "OrcamentoItemNecessarioResponse": {
        "type": "object",
        "properties": {
          "pecaInsumoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "servicoOsId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Filtro de óleo"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "PECA",
              "INSUMO"
            ],
            "example": "PECA"
          },
          "valorUnitario": {
            "type": "number",
            "format": "double",
            "example": 150
          },
          "quantidade": {
            "type": "integer",
            "example": 2
          },
          "valorTotal": {
            "type": "number",
            "format": "double",
            "example": 150
          }
        }
      },
      "OrcamentoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "ordemServicoId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "numeroOs": {
            "type": "string",
            "example": "OS-2026-0001"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "PRINCIPAL",
              "COMPLEMENTAR"
            ],
            "example": "PRINCIPAL"
          },
          "versao": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "DISPONIVEL",
              "APROVADO",
              "REPROVADO",
              "SUBSTITUIDO"
            ],
            "example": "DISPONIVEL"
          },
          "totalServicos": {
            "type": "number",
            "format": "double",
            "example": 150
          },
          "totalItens": {
            "type": "number",
            "format": "double",
            "example": 150
          },
          "totalGeral": {
            "type": "number",
            "format": "double",
            "example": 150
          },
          "servicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrcamentoServicoResponse"
            }
          },
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrcamentoItemNecessarioResponse"
            }
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          },
          "disponibilizadoEm": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-23T14:30:00"
          }
        }
      },
...
    }
  }
}